Reorganize sidebar index
